nvim: integrate with git
Add two git integration plugins to improve workflow productivity.

The first is vim-fugitive, which is a thin wrapper around the git CLI.
As such, its commands are similar to the commandline tool. This makes
knowledge more evergreen.

The second is vim-gitgutter, which notes modified hunks in the gutter
allowing for improved information density.