gentoo: update `@preserved-rebuild` and `@smart-live-rebuild` sets
This commit expands to update set to include `@preserved-rebuild` and
`@smart-live-rebuild` in addition to `@world`. Many times, `@world` will
necessitate an `emerge @preserved-rebuild` so to save time, update that
set.

Then it's easier to always update live rebuilds along with everything
else so we want to use `@live-rebuild`, but to avoid wasting time with
unnecessary rebuilds, we can use the `@smart-live-rebuild` provided by
`app-portage/smart-live-rebuild` package.