cli(nvim): use Lua APIs to configure LSP autocmds.
Another step towards a pure Lua configuration, we can use the autocmd
APIs to replace the embedded Vimscript. I've added a helper function for
adding autocmds to a group since the syntax was repetitive. It's a
curried function that returns a wrapper around
`vim.api.nvim_create_autocmd()`. It keeps the configuration a little
more terse.