cli(nvim): load a work profile if it exists
I'll put my work configuration at `~/.config/nvim/lua/config/work.lua`
so if it exists, we should try to load it.

One small note: To make it easier to navigate to each file, I've
modified the paths in the comments to be homedir rooted. `gf` works
better with this.