cli(nvim): move plugin configuration into `config/plugins/base.lua`
This large diff is just copy and pasting the plugin configuration in
`config/base.lua` into `config/plugins/base.lua` in the respective
plugin specs. I want to split this out further but this is a step in the
right direction.

With the plugin configuration in its own file, I've moved lazy.nvim's
`setup()` call into `init.lua` after setting the global Vim
configuration so it's easier to see where the plugins are loaded.