de(i3): start migration to systemd-based autostart.
Current state of affairs is a mix of systemd-based autostart and `exec`
statements in i3 config. It's hard to know exactly what programs are
running and whether it's part of the desktop environment or just an app.
I'd like all desktop environment programs to start in a more uniform
way.

To support this, we define the `i3-session.target`, started at the end
of the i3 config, which we will register programs that we only want in
an i3 session. This will be services such as the compositor and locker.
All of these programs will start when i3 starts the `i3-session.target`
and all the services can controlled together.

This commit only migrations picom as an example of what can be done.
Others will follow.