cli(nvim): use `toggleterm.nvim` instead of custom binding.
I've needed to use multiple terminals at once recently and the custom binding
hasn't been very ergonomic. Toggleterm looks to have better ergonomics and
lifecycle management.

The config is to use the same binding that I'm used to, disable the shading
since it looks weird to me, and have the terminal be larger than 20 lines. It's
such a small default.